Split generic types from the player structure.
